+++ modelo/src/entidades/performer.rs
@@ -1,7 +1,7 @@
 
 //representa Performer de la BD a un alto nivel
 pub struct Performer {
-    pub id: Option<u32>,
-    pub id_type: u32,
+    pub id: Option<i64>,
+    pub id_type: i64,
     pub name: String,
 }
