+++ modelo/src/entidades/person.rs
@@ -1,6 +1,6 @@
 //representa Person de la BD a un alto nivel
 pub struct Person {
-    pub id: Option<u32>,
+    pub id: Option<i64>,
     pub stage_name: String,
     pub real_name: String,
     pub birth_date: String,
