+++ modelo/src/entidades/rola.rs
@@ -1,11 +1,11 @@
 //representa Rola de la BD a un alto nivel
 pub struct Rola {
-    pub id: Option<u32>,
-    pub id_performer: u32,
-    pub id_album: u32,
+    pub id: Option<i64>,
+    pub id_performer: i64,
+    pub id_album: i64,
     pub path: String,
     pub title: String,
-    pub track: u32,
-    pub year: u32,
+    pub track: i64,
+    pub year: i64,
     pub genre: String,
 }
